refactor: move config file parser to another file
