+++ src/selector.rs
@@ -51,7 +51,8 @@ mod tests {
 }
 
 use crate::cmdline::Opt;
-use awsp::{get_aws_config_file_path, parse_config_file};
+
+use awsp::helper::file::config::{create_profile_config_map_from, get_aws_config_file_path};
 
 use dialoguer::{theme::ColorfulTheme, Select};
 use std::env;
@@ -142,7 +143,7 @@ fn display_selected() {
 
 fn profile_menu() {
     let location = get_aws_config_file_path().unwrap();
-    let config_file = parse_config_file(location.as_path()).unwrap();
+    let config_file = create_profile_config_map_from(location.as_path()).unwrap();
     let profile_list = to_key_list(&config_file);
     let profile_list = profile_list.as_slice();
     let default_profile = default_env("AWS_PROFILE");
